pyproject.toml
[tool.poetry]
name = "erp-backend"
version = "0.1.0"
description = "Microserviço destinado a manipulação dos dados tomadores e prestadores"
authors = ["ramonhveloso <ramon.veloso@inovebanco.com.br>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = "^0.30.6"
sqlalchemy = "^2.0.35"
pydantic = "^2.9.2"
passlib = "1.7.4"
jose = "^1.0.0"
python-jose = "^3.3.0"
pre-commit = "^3.8.0"
alembic = "^1.13.3"
autoflake = "^2.3.1"
mypy = "^1.11.2"
isort = "^5.13.2"
black = "^24.8.0"
stubs = "^1.0.0"
types-python-jose = "^3.3.4.20240106"
types-passlib = "^1.7.7.20240819"
email-validator = "^2.2.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9"
bcrypt = "3.2.0"
pytz = "^2024.2"
python-multipart = "^0.0.12"
pytest = "^8.3.3"
httpx = "^0.27.2"
pytest-asyncio = "^0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
known_first_party = ["app"]
default_section = "THIRDPARTY"
